fix(autopilot): architecture fixes for Loop integration (P1-P3)
- P1: Only update last_pod_key on successful completion to protect persistent sandbox resume chain
- P1: Add approval timeout detection in LoopScheduler (CheckApprovalTimeouts)
- P2: Map max_iterations to dedicated terminal phase instead of generic "stopped"
- P2: Cap control process stdout/stderr at 10MB to prevent Runner OOM
- P3: Add 30s timeout to wg.Wait() in Stop() to prevent indefinite blocking
- Refactor: Split loop_orchestrator.go into 5 SRP files (trigger/start/completion/monitoring/notify)
- Fix: Distinguish truncated output from process failure in control_runner logs
