+++ runner/internal/autopilot/command_executor.go
@@ -4,9 +4,40 @@ package autopilot
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"os/exec"
 )
+
+// maxOutputSize is the maximum bytes captured from stdout/stderr of a control
+// process. Claude CLI with --output-format json can produce very large output
+// (full tool_use chains, code blocks, etc.). Capping prevents Runner OOM.
+const maxOutputSize = 10 * 1024 * 1024 // 10 MB
+
+// errOutputTruncated is returned when command output exceeds maxOutputSize.
+var errOutputTruncated = errors.New("command output truncated: exceeded maximum size")
+
+// limitedWriter is a bytes.Buffer wrapper that stops writing after a size limit.
+// Once the limit is reached, subsequent writes are silently discarded.
+type limitedWriter struct {
+	buf   bytes.Buffer
+	limit int
+}
+
+func (w *limitedWriter) Write(p []byte) (int, error) {
+	remaining := w.limit - w.buf.Len()
+	if remaining <= 0 {
+		// Silently discard — report full len to prevent exec from erroring
+		return len(p), nil
+	}
+	if len(p) > remaining {
+		p = p[:remaining]
+	}
+	return w.buf.Write(p)
+}
+
+func (w *limitedWriter) Bytes() []byte { return w.buf.Bytes() }
+func (w *limitedWriter) Len() int      { return w.buf.Len() }
 
 // CommandExecutor abstracts command execution for testability and extensibility.
 // This follows the Dependency Inversion Principle (DIP).
@@ -25,13 +56,15 @@ func NewDefaultCommandExecutor() *DefaultCommandExecutor {
 }
 
 // Execute runs a command using os/exec.
+// stdout and stderr are each capped at maxOutputSize (10 MB) to prevent OOM.
 func (e *DefaultCommandExecutor) Execute(ctx context.Context, name string, args []string, workDir string) ([]byte, []byte, error) {
 	cmd := exec.CommandContext(ctx, name, args...)
 	cmd.Dir = workDir
 
-	var stdout, stderr bytes.Buffer
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
+	stdout := &limitedWriter{limit: maxOutputSize}
+	stderr := &limitedWriter{limit: maxOutputSize}
+	cmd.Stdout = stdout
+	cmd.Stderr = stderr
 
 	err := cmd.Run()
 	if err != nil {
@@ -39,6 +72,11 @@ func (e *DefaultCommandExecutor) Execute(ctx context.Context, name string, args 
 			return stdout.Bytes(), stderr.Bytes(), fmt.Errorf("command timed out: %w", ctx.Err())
 		}
 		return stdout.Bytes(), stderr.Bytes(), fmt.Errorf("command failed: %w", err)
+	}
+
+	// Warn caller if output was truncated
+	if stdout.Len() >= maxOutputSize || stderr.Len() >= maxOutputSize {
+		return stdout.Bytes(), stderr.Bytes(), errOutputTruncated
 	}
 
 	return stdout.Bytes(), stderr.Bytes(), nil
